Ideal gas law

Calculate the density, pressure or temperature or air from the other two variables.

IdealGas(p, t, rho, R = 287.058)

Arguments

p
pressure

t
temperature

rho
density

R
gas constant for air

Value

Pressure, temperature or density, according to which variable was not passed as an argument.

Details

Is important to take note of the units in which each variable is provided. With the default value of R, pressure should be passed in Pascals, temperature in Kelvin and density in \(kg/m^3\). The result will be also in those units.

See also

Other meteorology functions: Derivate, EOF, FitQsWave, GeostrophicWind, WaveFlux

Examples

IdealGas(1013*100, 20 + 273.15)

#> [1] 1.203788

IdealGas(1013*100, rho = 1.15) - 273.15

#> [1] 33.71118