Quasi stationary waves

Quasi stationary waves

FitQsWave(y, k = 1)

BuildQsField(x, amplitude, phase, k)

Arguments

y
numeric vector to transform

k
numeric vector of wave numbers to keep or wave numbers to compute

x
numeric vector of locations (in radians)

amplitude
numeric vector of amplitudes

phase
numeric vector of phases

Value

a named list of amplitude, phase and wave number. BuildQsField returns a vector.

Details

FitQsWave uses fft to make a fourier transform of the data and then returns a list of parameters for each wave number kept. The amplitude (A), phase (\(\phi\)) and wave number (k) satisfy: $$y = \sum A cos((x - \phi)k)$$ The phase is calculated so that it lies between 0 and \(2\pi/k\) so it represents the location (in radians) of the first maximum of each wave number. BuildQsField is, in a way, the opposite. It computes the the above mentioned sum but for only the selected wave numbers.

See also

Other meteorology functions: Derivate, EOF, GeostrophicWind, IdealGas, WaveFlux