Relief Shading

geom_relief() simulates shading caused by relief. Can be useful when plotting topographic data because relief shading might give a more intuitive impression of the shape of the terrain than contour lines or mapping height to color. geom_shadow() projects shadows.

geom_relief(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., sun.angle = 60, raster = TRUE,
  interpolate = TRUE, shadow = FALSE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

geom_shadow(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., sun.angle = 60, range = c(0, 1),
  skip = 0, raster = TRUE, interpolate = TRUE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data.

stat
The statistical transformation to use on the data for this layer, as a string.

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

sun.angle
angle from which the sun is shining, in degrees counterclockwise from 12 o' clock

raster
if TRUE (the default), uses ggplot2::geom_raster, if FALSE, uses ggplot2::geom_tile.

interpolate
If TRUE interpolate linearly, if FALSE (the default) don't interpolate.

shadow
if TRUE, adds also a layer of geom_shadow()

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

range
transparency range for shadows

skip
data points to skip when casting shadows

Details

light and dark must be valid colours determining the light and dark shading (defaults to "white" and "gray20", respectively).

Aesthetics

geom_relief() and geom_shadow() understands the following aesthetics (required aesthetics are in bold)

x

y

z

light

dark

sun.angle

See also

Examples

# NOT RUN {
library(ggplot2)
ggplot(reshape2::melt(volcano), aes(Var1, Var2)) +
      geom_relief(aes(z = value))
# }