Methylation data in chromosome 1.

A dataset containing the methylation data in chromosome 1. Data for this probe are downloaded from site: http://www.neuroepigenomics.org/methylomedb/download.html. We used 4 control samples: Control1 AC, Control2 AC, Control3 AC and Control4 AC and 4 disease samples: SCZ1 AC, SCZ2 AC, SCZ3 AC, SCZ4 AC. After downloading we summed all results on the same position and chromosome separately from control and disease samples. The variables are as follows:

data(schizophrenia)

Format

A data frame with 4 116 194 rows and 7 variables

Details

chr. name of chromosome where methylation data was collected (only chromosome 1.)

poz. position of chromosome where methylation data was collected

no. number of reads per position

meth. number of reads with methylated cytosine

unmeth. number of reads with unmethylated cytosine

meth.rate. methylation.rate that is ratio meth./(meth. + unmeth.)

category indicates if data are from control samples ('control' value) or from disease samples ('disease' value)