Bilinear inteprolation

Uses fields::interp.surface to interpolate values defined in a bidimensional grid with bilinear interpolation.

Interpolate(formula, x.out, y.out, data = NULL, grid = TRUE)

Arguments

formula
a formula indicating dependent and independent variables (see Details)

x.out, y.out
x and y values where to interpolate (see Details)

data
optional data.frame with the data

grid
logical indicating if x.out and y.out define a regular grid.

Value

A data.frame with interpolated values and locations

Details

formula must be of the form VAR1 | VAR2 ~ X + Y where VAR1, VAR2, etc... are the names of the variables to interpolate and X and Y the names of the x and y values, respectively. It is also possible to pass only values of x, in which case, regular linear interpolation is performed and y.out, if exists, is ignored with a warning.

If grid = TRUE, x.out and y.out must define the values of a regular grid. If grid = FALSE, they define the locations where to interpolate.

Examples

library(data.table)
data(geopotential)
geopotential <- geopotential[date == date[1]]
# new grid
x.out <- seq(0, 360, by = 10)
y.out <- seq(-90, 0, by = 10)

# Interpolate values to a new grid
interpolated <- geopotential[, Interpolate(gh ~ lon + lat, x.out, y.out)]

# Add values to an existing grid
geopotential[, gh.new := Interpolate(gh ~ lon + lat, lon, lat,
                                     data = interpolated, grid = FALSE)$gh]

#>         lon   lat lev       gh       date   gh.new
#>    1:   0.0 -22.5 700 3163.839 1990-01-01       NA
#>    2:   2.5 -22.5 700 3162.516 1990-01-01       NA
#>    3:   5.0 -22.5 700 3162.226 1990-01-01       NA
#>    4:   7.5 -22.5 700 3162.323 1990-01-01       NA
#>    5:  10.0 -22.5 700 3163.097 1990-01-01       NA
#>   ---                                             
#> 4028: 347.5 -90.0 700 2715.936 1990-01-01 2715.936
#> 4029: 350.0 -90.0 700 2715.936 1990-01-01       NA
#> 4030: 352.5 -90.0 700 2715.936 1990-01-01       NA
#> 4031: 355.0 -90.0 700 2715.936 1990-01-01       NA
#> 4032: 357.5 -90.0 700 2715.936 1990-01-01       NA

# Interpolate multiple values
geopotential[, c("u", "v") := GeostrophicWind(gh, lon, lat)]

#>         lon   lat lev       gh       date   gh.new  u          v
#>    1:   0.0 -22.5 700 3163.839 1990-01-01       NA NA  1.0807091
#>    2:   2.5 -22.5 700 3162.516 1990-01-01       NA NA  0.5513294
#>    3:   5.0 -22.5 700 3162.226 1990-01-01       NA NA  0.0661829
#>    4:   7.5 -22.5 700 3162.323 1990-01-01       NA NA -0.2976979
#>    5:  10.0 -22.5 700 3163.097 1990-01-01       NA NA -0.7498781
#>   ---                                                           
#> 4028: 347.5 -90.0 700 2715.936 1990-01-01 2715.936 NA  0.0000000
#> 4029: 350.0 -90.0 700 2715.936 1990-01-01       NA NA  0.0000000
#> 4030: 352.5 -90.0 700 2715.936 1990-01-01       NA NA  0.0000000
#> 4031: 355.0 -90.0 700 2715.936 1990-01-01       NA NA  0.0000000
#> 4032: 357.5 -90.0 700 2715.936 1990-01-01       NA NA  0.0000000

interpolated <- geopotential[, Interpolate(u | v ~ lon + lat, x.out, y.out)]