Reference arrow for magnitude scales

Draws a reference arrow. Highly experimental.

guide_vector(
  title = ggplot2::waiver(),
  title.position = NULL,
  title.theme = NULL,
  title.hjust = NULL,
  title.vjust = NULL,
  label = TRUE,
  label.position = NULL,
  label.theme = NULL,
  label.hjust = NULL,
  label.vjust = NULL,
  keywidth = NULL,
  keyheight = NULL,
  direction = NULL,
  default.unit = "cm",
  override.aes = list(),
  nrow = NULL,
  ncol = NULL,
  byrow = FALSE,
  reverse = FALSE,
  order = 0,
  ...
)

Arguments

title
A character string or expression indicating a title of guide. If NULL, the title is not shown. By default (waiver()), the name of the scale object or the name specified in labs() is used for the title.

title.position
A character string indicating the position of a title. One of "top" (default for a vertical guide), "bottom", "left" (default for a horizontal guide), or "right."

title.theme
A theme object for rendering the title text. Usually the object of element_text() is expected. By default, the theme is specified by legend.title in theme() or theme.

title.hjust
A number specifying horizontal justification of the title text.

title.vjust
A number specifying vertical justification of the title text.

label
logical. If TRUE then the labels are drawn. If FALSE then the labels are invisible.

label.position
A character string indicating the position of a label. One of "top", "bottom" (default for horizontal guide), "left", or "right" (default for vertical guide).

label.theme
A theme object for rendering the label text. Usually the object of element_text() is expected. By default, the theme is specified by legend.text in theme().

label.hjust
A numeric specifying horizontal justification of the label text.

label.vjust
A numeric specifying vertical justification of the label text.

keywidth
A numeric or a grid::unit() object specifying the width of the legend key. Default value is legend.key.width or legend.key.size in theme().

keyheight
A numeric or a grid::unit() object specifying the height of the legend key. Default value is legend.key.height or legend.key.size in theme().

direction
A character string indicating the direction of the guide. One of "horizontal" or "vertical."

default.unit
A character string indicating grid::unit() for keywidth and keyheight.

override.aes
A list specifying aesthetic parameters of legend key. See details and examples.

nrow
The desired number of rows of legends.

ncol
The desired number of column of legends.

byrow
logical. If FALSE (the default) the legend-matrix is filled by columns, otherwise the legend-matrix is filled by rows.

reverse
logical. If TRUE the order of legends is reversed.

order
positive integer less than 99 that specifies the order of this guide among multiple guides. This controls the order in which multiple guides are displayed, not the contents of the guide itself. If 0 (default), the order is determined by a secret algorithm.

...
ignored.

See also